Login, register fix
