--- aMovieDatabase/core/templates/login.html
@@ -1,24 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-
-{% block content %}
-  <div class="container mt-5">
-    <div class="row justify-content-center">
-      <div class="col-lg-4 center">
-        <form method="post" action="{% url 'login' %}" class="form-signin">
-          {% csrf_token %}
-          <h1 class="h3 mb-3 font-weight-normal text-center">Please sign in</h1>
-          <div class="form-group">
-            {{ form.username.label_tag }}
-            {{ form.username }}
-          </div>
-          <div class="form-group clearfix">
-            {{ form.password.label_tag }}
-            {{ form.password }}
-          </div>
-          <input type="submit" value="Log in" class="btn btn-primary btn-block">
-        </form>
-      </div>
-    </div>
-  </div>
-{% endblock %}
