--- aMovieDatabase/core/templates/register.html
@@ -1,36 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-
-{% block content %}
-<div class="container mt-5">
- <div class="d-flex align-items-center h-100">
-  <div class="mx-auto w-50">
-    <form method="post" action="{% url 'register' %}" class="form-signin">
-      {% csrf_token %}
-      <h1 class="h3 mb-3 font-weight-normal text-center">Please register</h1>
-      <div class="form-group">
-        {{ form.first_name.label_tag }}
-        {{ form.first_name }}
-      </div>
-      <div class="form-group">
-        {{ form.last_name.label_tag }}
-        {{ form.last_name }}
-      </div>
-      <div class="form-group">
-        {{ form.email.label_tag }}
-        {{ form.email }}
-      </div>
-      <div class="form-group clearfix">
-        {{ form.password1.label_tag }}
-        {{ form.password1 }}
-      </div>
-      <div class="form-group clearfix">
-        {{ form.password2.label_tag }}
-        {{ form.password2 }}
-      </div>
-      <input type="submit" value="Register" class="btn btn-primary btn-block text-center">
-    </form>
-  </div>
-</div>
- </div>
-  {% endblock %}
